+++ desc/objectives/objective_funs.py
@@ -435,7 +435,7 @@ class ObjectiveFunction(IOAble):
                 pass
 
     @execute_on_cpu
-    def build(self, use_jit=None, verbose=1):
+    def build(self, use_jit=None, verbose=1):  # noqa: C901
         """Build the objective.
 
         Parameters
@@ -511,15 +511,18 @@ class ObjectiveFunction(IOAble):
             else:
                 self._deriv_mode = "blocked"
 
+        rev_objs = [
+            o.__class__.__name__ for o in self.objectives if o._deriv_mode == "rev"
+        ]
         warnif(
-            any((obj._deriv_mode == "rev") for obj in self.objectives)
-            and self._deriv_mode == "batched",
+            len(rev_objs) > 0 and self._deriv_mode == "batched",
             UserWarning,
-            "Using batched derivative mode, which performs fwd mode "
-            "differentiation of the entire objective, but one or more of the"
-            ' sub-objectives have deriv_mode="rev". It is recommended to use '
-            'deriv_mode="blocked", as these objective performances may suffer '
-            "in fwd mode, or may not even support reverse mode at all.",
+            "'batched' deriv_mode differentiates the whole ObjectiveFunction in "
+            "forward mode, but these sub-objectives are set to use reverse mode "
+            "(either automatically, from their input/output sizes, or by user): "
+            f"{rev_objs}. In forward mode these may under-perform, or may not "
+            "work at all. Consider 'blocked' deriv_mode. See the sub-objective "
+            "docstrings for details.",
         )
 
         errorif(
@@ -557,6 +560,12 @@ class ObjectiveFunction(IOAble):
 
         timer.stop("Objective build")
         if verbose > 1:
+            print(f"{self.name} deriv_mode : {self._deriv_mode}")
+            if self._deriv_mode == "batched":
+                print(f"{self.name} jac_chunk_size: {self._jac_chunk_size}")
+            else:
+                for o in self.objectives:
+                    print(f"{o.name} jac_chunk_size: {o._jac_chunk_size}")
             timer.disp("Objective build")
 
     def _set_things(self, things=None):